requirements.txt
selenium
requests
Pillow
webdriver-manager
split-folders
tensorflow
